@@ -3,16 +3,11 @@ apply plugin: 'war'
 apply plugin: 'eclipse-wtp'
 
 ext {
-	javaVersion = JavaVersion.VERSION_1_8
+	javaVersion = JavaVersion.VERSION_21
 }
 
 repositories {
 	mavenCentral()
-
-	maven {
-		// for jasper reports
-		url 'https://jaspersoft.jfrog.io/jaspersoft/third-party-ce-artifacts/'
-	}
 
 	if (project.hasProperty('iPLAssMavenRepoUrl')) {
 		maven {
