@@ -22,8 +22,8 @@ package samples.ec01.bean;
 
 import java.io.Serializable;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 
 import org.iplass.mtp.web.template.TemplateUtil;
 
