@@ -22,9 +22,9 @@ package samples.ec01.bean;
 
 import java.io.Serializable;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
 
 import org.iplass.mtp.auth.User;
 import org.iplass.mtp.web.template.TemplateUtil;
